task-4/requirements.txt
streamlit
transformers
google-generativeai 
langchain_google_genai
langchain
langchain_community
langchain_core
python-dotenv